+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # dickinson
 
+  * Fix bug in `:type`
+
 # 1.0.0.1
 
   * Fix bug when calling `:flatten` on tuples.