+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # dickinson
 
+  * Fix bug in `:type`
+  * Add builtins `oulipo`, `allCaps`, `capitalize`, and `titlecase`
+
 # 1.0.0.1
 
   * Fix bug when calling `:flatten` on tuples.